Refactor string assertions
